requirements.txt
PyPDF2
python-pptx
python-docx
requests
python-dotenv
